wip: infer rows
