+++ polars/polars-core/Cargo.toml
@@ -36,7 +36,7 @@ plain_fmt = ["prettytable-rs"]
 sort_multiple = []
 # create from row values
 # and include pivot operation
-rows = []
+rows = ["indexmap"]
 # dont use this
 private = []
 
@@ -160,6 +160,7 @@ serde = { version = "1", features = ["derive"], optional = true }
 serde_json = { version = "1.0", optional = true }
 thiserror = "^1.0"
 unsafe_unwrap = "^0.1.0"
+indexmap = { version = "^1.6", optional = true }
 
 [dependencies.arrow]
 package = "arrow2"
